Drastical speed up of 2024/19.
Store the available towels in a Hash of length => Set of towels with
that length. Runtime is now reduced by over 80%. The majority of this
comes from the simplified lookup of available patterns, but some of it
is due to the fact that the number of times the remaining pattern needs
to be cut is reduced.

Since this is about twice as fast as the alternative solution, that
solution is now redundant and removed.